web/requirements.txt
Django==2.2
Pillow==5.4
django-ckeditor==5.6.1
libsass==0.18.0
django-sass-processor==0.7.2
django-cities-light==3.5
python-slugify==3.0
psycopg2>=2.7,<3.0
